+++ lib/carrierwave/uploader/content_type_whitelist.rb
@@ -33,14 +33,19 @@ module CarrierWave
     private
 
       def check_content_type_whitelist!(new_file)
+        return unless content_type_whitelist
+
         content_type = new_file.content_type
-        if content_type_whitelist && !whitelisted_content_type?(content_type)
+        if !whitelisted_content_type?(content_type)
           raise CarrierWave::IntegrityError, I18n.translate(:"errors.messages.content_type_whitelist_error", content_type: content_type, allowed_types: Array(content_type_whitelist).join(", "))
         end
       end
 
       def whitelisted_content_type?(content_type)
-        Array(content_type_whitelist).any? { |item| content_type =~ /#{item}/ }
+        Array(content_type_whitelist).any? do |item|
+          item = Regexp.quote(item) if item.class != Regexp
+          content_type =~ /#{item}/
+        end
       end
 
     end # ContentTypeWhitelist