+++ lib/carrierwave/uploader/extension_blacklist.rb
@@ -37,8 +37,10 @@ module CarrierWave
     private
 
       def check_extension_blacklist!(new_file)
+        return unless extension_blacklist
+
         extension = new_file.extension.to_s
-        if extension_blacklist && blacklisted_extension?(extension)
+        if blacklisted_extension?(extension)
           raise CarrierWave::IntegrityError, I18n.translate(:"errors.messages.extension_blacklist_error", extension: new_file.extension.inspect, prohibited_types: Array(extension_blacklist).join(", "))
         end
       end