+++ spec/downloader/base_spec.rb
@@ -26,14 +26,15 @@ describe CarrierWave::Downloader::Base do
     end
   end
 
-  context "with a URL with internationalized domain name" do
-    let(:uri) { "http://ドメイン名例.jp/#{CGI.escape(filename)}" }
+  context "with equal and colons in the query path" do
+    let(:query) { 'test=query&with=equal&before=colon:param' }
+    let(:uri) { "https://example.com/#{filename}?#{query}" }
     before do
-      stub_request(:get, 'http://xn--eckwd4c7cu47r2wf.jp/test.jpg').to_return(body: file)
+      stub_request(:get, uri).to_return(body: file)
     end
 
-    it "converts to Punycode URI" do
-      expect(subject.process_uri(uri).to_s).to eq 'http://xn--eckwd4c7cu47r2wf.jp/test.jpg'
+    it "leaves colon in resulting URI" do
+      expect(subject.process_uri(uri).query).to eq query
     end
 
     it "downloads a file" do
@@ -61,10 +62,6 @@ describe CarrierWave::Downloader::Base do
     end
   end
 
-  it "raises an error when trying to download a local file" do
-    expect { subject.download('/etc/passwd') }.to raise_error(CarrierWave::DownloadError)
-  end
-
   context "with missing file" do
     before do
       stub_request(:get, uri).to_return(status: 404)
@@ -76,6 +73,20 @@ describe CarrierWave::Downloader::Base do
 
     it "doesn't obscure original exception message" do
       expect { subject.download(uri) }.to raise_error(CarrierWave::DownloadError, /could not download file: 404/)
+    end
+  end
+
+  context "with server error" do
+    before do
+      stub_request(:get, uri).to_return(status: 500)
+    end
+
+    it "raises an error when trying to download" do
+      expect{ subject.download(uri) }.to raise_error(CarrierWave::DownloadError)
+    end
+
+    it "doesn't obscure original exception message" do
+      expect { subject.download(uri) }.to raise_error(CarrierWave::DownloadError, /could not download file: 500/)
     end
   end
 
@@ -95,7 +106,7 @@ describe CarrierWave::Downloader::Base do
     before do
       stub_request(:get, uri).
         to_return(status: 301, body: "Redirecting", headers: { "Location" => another_uri })
-      stub_request(:get, another_uri).to_return(body: file)
+      stub_request(:get, /redirected.jpg/).to_return(body: file)
     end
 
     it "retrieves redirected file" do
@@ -107,7 +118,31 @@ describe CarrierWave::Downloader::Base do
     end
   end
 
+  context "with SSRF prevention" do
+    before do
+      stub_request(:get, 'http://169.254.169.254/').to_return(body: file)
+      stub_request(:get, 'http://127.0.0.1/').to_return(body: file)
+    end
+
+    it "prevents accessing local files" do
+      expect { subject.download('/etc/passwd') }.to raise_error(CarrierWave::DownloadError)
+      expect { subject.download('file:///etc/passwd') }.to raise_error(CarrierWave::DownloadError)
+    end
+
+    it "prevents accessing internal addresses" do
+      expect { uploader.download!("http://169.254.169.254/") }.to raise_error CarrierWave::DownloadError
+      expect { uploader.download!("http://lvh.me/") }.to raise_error CarrierWave::DownloadError
+    end
+  end
+
   describe '#process_uri' do
+    it "converts a URL with internationalized domain name to Punycode URI" do
+      uri = "http://ドメイン名例.jp/#{CGI.escape(filename)}"
+      processed = subject.process_uri(uri)
+      expect(processed.class).to eq(URI::HTTP)
+      expect(processed.to_s).to eq 'http://xn--eckwd4c7cu47r2wf.jp/test.jpg'
+    end
+
     it "parses but not escape already escaped uris" do
       uri = 'http://example.com/%5B.jpg'
       processed = subject.process_uri(uri)
@@ -129,6 +164,13 @@ describe CarrierWave::Downloader::Base do
       expect(processed.to_s).to eq('http://example.com/%20%25%5B%5D.jpg')
     end
 
+    it "parses but not escape uris with query-string characters representing urls not needing escaping " do
+      uri = 'http://example.com/?src0=https%3A%2F%2Fi.vimeocdn.com%2Fvideo%2F1234_1280x720.jpg'
+      processed = subject.process_uri(uri)
+      expect(processed.class).to eq(URI::HTTP)
+      expect(processed.to_s).to eq(uri)
+    end
+
     it "escapes and parse brackets in uri paths without harming the query string" do
       uri = 'http://example.com/].jpg?test[]'
       processed = subject.process_uri(uri)
@@ -136,9 +178,35 @@ describe CarrierWave::Downloader::Base do
       expect(processed.to_s).to eq('http://example.com/%5D.jpg?test[]')
     end
 
+    it "escapes and parse unescaped characters in path" do
+      uri = 'http://example.com/あああ.jpg'
+      processed = subject.process_uri(uri)
+      expect(processed.class).to eq(URI::HTTP)
+      expect(processed.to_s).to eq('http://example.com/%E3%81%82%E3%81%82%E3%81%82.jpg')
+    end
+
+    it "escapes and parse unescaped characters in query string" do
+      uri = 'http://example.com/?q=あああ'
+      processed = subject.process_uri(uri)
+      expect(processed.class).to eq(URI::HTTP)
+      expect(processed.to_s).to eq('http://example.com/?q=%E3%81%82%E3%81%82%E3%81%82')
+    end
+
     it "throws an exception on bad uris" do
       uri = '~http:'
       expect { subject.process_uri(uri) }.to raise_error(CarrierWave::DownloadError)
+    end
+  end
+
+  describe "#skip_ssrf_protection?" do
+    let(:uri) { 'http://localhost/test.jpg' }
+    before do
+      WebMock.stub_request(:get, uri).to_return(body: file)
+      allow(subject).to receive(:skip_ssrf_protection?).and_return(true)
+    end
+
+    it "allows local request to be made" do
+      expect(subject.download(uri).read).to eq 'this is stuff'
     end
   end
 end