+++ spec/storage/fog_helper.rb
@@ -97,6 +97,9 @@ end
             if @provider == 'AWS'
               expect(@storage.connection).to receive(:copy_object)
                                               .with(anything, anything, anything, anything, { "Content-Type"=>file.content_type, "x-amz-acl"=>"public-read" }).and_call_original
+            elsif @provider == 'Google'
+              expect(@storage.connection).to receive(:copy_object)
+                                              .with(anything, anything, anything, anything, { "Content-Type"=>file.content_type, destination_predefined_acl: "publicRead" }).and_call_original
             else
               expect(@storage.connection).to receive(:copy_object)
                                               .with(anything, anything, anything, anything, { "Content-Type"=>file.content_type }).and_call_original